+++ lib/model/berry.g.dart
@@ -1,0 +1,17 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'berry.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+Berry _$BerryFromJson(Map<String, dynamic> json) => Berry(
+      json['name'] as String,
+      json['url'] as String,
+    );
+
+Map<String, dynamic> _$BerryToJson(Berry instance) => <String, dynamic>{
+      'name': instance.name,
+      'url': instance.url,
+    };