+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+### 28-12-2021
+- [ ] make api crawler poke api
+- [ ] make screen search pokemon
+- [x] make screen list berry
+
 ### 27-12-2021
 - [x] init project
 - [x] fetch data